ai_engine/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
pypdf
python-docx
